// webui/src/lib/gen/models/nIMBUSClassificationRequest.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * DESDEO (fast)API
 * A rest API for the DESDEO framework.
 * OpenAPI spec version: 0.1.0
 */
import type { NIMBUSClassificationRequestSessionId } from './nIMBUSClassificationRequestSessionId';
import type { NIMBUSClassificationRequestParentStateId } from './nIMBUSClassificationRequestParentStateId';
import type { NIMBUSClassificationRequestScalarizationOptions } from './nIMBUSClassificationRequestScalarizationOptions';
import type { NIMBUSClassificationRequestSolver } from './nIMBUSClassificationRequestSolver';
import type { NIMBUSClassificationRequestSolverOptions } from './nIMBUSClassificationRequestSolverOptions';
import type { ReferencePoint } from './referencePoint';
import type { NIMBUSClassificationRequestCurrentObjectives } from './nIMBUSClassificationRequestCurrentObjectives';
import type { NIMBUSClassificationRequestNumDesired } from './nIMBUSClassificationRequestNumDesired';

/**
 * Model of the request to the nimbus method.
 */
export interface NIMBUSClassificationRequest {
	problem_id: number;
	session_id?: NIMBUSClassificationRequestSessionId;
	parent_state_id?: NIMBUSClassificationRequestParentStateId;
	scalarization_options?: NIMBUSClassificationRequestScalarizationOptions;
	solver?: NIMBUSClassificationRequestSolver;
	solver_options?: NIMBUSClassificationRequestSolverOptions;
	preference?: ReferencePoint;
	/** The objectives used for iteration. */
	current_objectives: NIMBUSClassificationRequestCurrentObjectives;
	num_desired?: NIMBUSClassificationRequestNumDesired;
}