// webui/src/lib/gen/models/NIMBUSInitializationRequest.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * DESDEO (fast)API
 * A rest API for the DESDEO framework.
 * OpenAPI spec version: 0.1.0
 */
import type { NIMBUSInitializationRequestScalarizationOptions } from './NIMBUSInitializationRequestScalarizationOptions';
import type { NIMBUSInitializationRequestSolverOptions } from './NIMBUSInitializationRequestSolverOptions';
import type { ReferencePoint } from './ReferencePoint';
import type { SolutionInfo } from './SolutionInfo';

/**
 * Model of the request to the nimbus method.
 */
export interface NIMBUSInitializationRequest {
	problem_id: number;
	session_id?: number | null;
	parent_state_id?: number | null;
	starting_point?: ReferencePoint | SolutionInfo | null;
	scalarization_options?: NIMBUSInitializationRequestScalarizationOptions;
	solver?: string | null;
	solver_options?: NIMBUSInitializationRequestSolverOptions;
}