// webui/src/lib/gen/models/nIMBUSInitializationRequest.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * DESDEO (fast)API
 * A rest API for the DESDEO framework.
 * OpenAPI spec version: 0.1.0
 */
import type { NIMBUSInitializationRequestSessionId } from './nIMBUSInitializationRequestSessionId';
import type { NIMBUSInitializationRequestParentStateId } from './nIMBUSInitializationRequestParentStateId';
import type { NIMBUSInitializationRequestStartingPoint } from './nIMBUSInitializationRequestStartingPoint';
import type { NIMBUSInitializationRequestScalarizationOptions } from './nIMBUSInitializationRequestScalarizationOptions';
import type { NIMBUSInitializationRequestSolver } from './nIMBUSInitializationRequestSolver';
import type { NIMBUSInitializationRequestSolverOptions } from './nIMBUSInitializationRequestSolverOptions';

/**
 * Model of the request to the nimbus method.
 */
export interface NIMBUSInitializationRequest {
	problem_id: number;
	session_id?: NIMBUSInitializationRequestSessionId;
	parent_state_id?: NIMBUSInitializationRequestParentStateId;
	starting_point?: NIMBUSInitializationRequestStartingPoint;
	scalarization_options?: NIMBUSInitializationRequestScalarizationOptions;
	solver?: NIMBUSInitializationRequestSolver;
	solver_options?: NIMBUSInitializationRequestSolverOptions;
}