// webui/src/lib/gen/models/NIMBUSFinalizeRequest.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * DESDEO (fast)API
 * A rest API for the DESDEO framework.
 * OpenAPI spec version: 0.1.0
 */
import type { SolutionInfo } from './SolutionInfo';

/**
 * Request model for finalizing the NIMBUS procedure.
 */
export interface NIMBUSFinalizeRequest {
	problem_id: number;
	session_id?: number | null;
	parent_state_id?: number | null;
	solution_info: SolutionInfo;
}