// webui/src/lib/gen/models/intermediateSolutionRequest.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * DESDEO (fast)API
 * A rest API for the DESDEO framework.
 * OpenAPI spec version: 0.1.0
 */
import type { IntermediateSolutionRequestSessionId } from './intermediateSolutionRequestSessionId';
import type { IntermediateSolutionRequestParentStateId } from './intermediateSolutionRequestParentStateId';
import type { IntermediateSolutionRequestContext } from './intermediateSolutionRequestContext';
import type { IntermediateSolutionRequestScalarizationOptions } from './intermediateSolutionRequestScalarizationOptions';
import type { IntermediateSolutionRequestSolver } from './intermediateSolutionRequestSolver';
import type { IntermediateSolutionRequestSolverOptions } from './intermediateSolutionRequestSolverOptions';
import type { IntermediateSolutionRequestNumDesired } from './intermediateSolutionRequestNumDesired';
import type { SolutionInfo } from './solutionInfo';

/**
 * Model of the request to solve intermediate solutions between two solutions.
 */
export interface IntermediateSolutionRequest {
	problem_id: number;
	session_id?: IntermediateSolutionRequestSessionId;
	parent_state_id?: IntermediateSolutionRequestParentStateId;
	context?: IntermediateSolutionRequestContext;
	scalarization_options?: IntermediateSolutionRequestScalarizationOptions;
	solver?: IntermediateSolutionRequestSolver;
	solver_options?: IntermediateSolutionRequestSolverOptions;
	num_desired?: IntermediateSolutionRequestNumDesired;
	reference_solution_1: SolutionInfo;
	reference_solution_2: SolutionInfo;
}