// webui/src/lib/gen/models/NIMBUSMultiplierRequest.ts
/**
 * Generated by orval v8.0.2 🍺
 * Do not edit manually.
 * DESDEO (fast)API
 * A rest API for the DESDEO framework.
 * OpenAPI spec version: 0.1.0
 */

/**
 * Request to get Lagrange multipliers from the solutions in a state.
 */
export interface NIMBUSMultiplierRequest {
	state_id: number;
	objective_symbols?: string[] | null;
}